Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,7 +2,7 @@
 int main() {
     float a,b,c;
     scanf("%f %f %f",&a,&b,&c);
-    printf("Average: %.2f",a+b+c/(float)3);
+    printf("Average: %.2f",a+b+c/3.00);
  
     return 0;
 }
